Move `app/javascript/hooks` to `app/javascript/mastodon/hooks`
