+++ app/javascript/mastodon/hooks/useHovering.ts
