+++ app/javascript/mastodon/hooks/useLinks.ts
