+++ app/javascript/mastodon/hooks/useRenderSignal.ts
