+++ app/javascript/mastodon/hooks/useSearchParam.ts
