+++ app/javascript/mastodon/hooks/useSelectableClick.ts
