+++ app/javascript/mastodon/hooks/useTimeout.ts
